django ckeditor eklendi

+++ Python/django/django2/django2/post/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render,HttpResponse,get_object_or_404,HttpResponseRedirect,redirect
+from django.shortcuts import render,HttpResponse,get_object_or_404,HttpResponseRedirect,redirect,Http404
 from .models import Post
 from .forms import PostForm
 from django.contrib import messages
@@ -18,6 +18,8 @@ def post_detail(request,id):
     return render(request,"post/detail.html",context)
 
 def post_create(request):
+    if not request.user.is_authenticated:
+        return Http404
 #    form = PostForm()   
 #    if request.method == "POST":
         # Formdan gelen bilgileri kaydet
@@ -42,6 +44,8 @@ def post_create(request):
     return render(request, "post/form.html",context)
  
 def post_update(request, id):
+    if not request.user.is_authenticated:
+        return Http404
     post = get_object_or_404(Post, id = id)
     form = PostForm(request.POST or None,request.FILES or None,instance=post)
     if form.is_valid():
@@ -53,6 +57,8 @@ def post_update(request, id):
     }
     return render(request, "post/form.html",context)
 def post_delete(request,id):
+    if not request.user.is_authenticated:
+        return Http404
     post = get_object_or_404(Post, id = id)
     post.delete()
     return redirect("post:index")
